Suppliers SKO Units and Selling Outer (CSER-10)
As i spoke to staff over here and they still don't understand what they
are booking in when it arrives and also export guys missing the
informations on how we sell the product if it coimes in 1 unit per sko
and we decide to sell on wen in 3's
I would recommend to add onto suppliers part and inventory Selling
outer section.
So when it comes to ordering we will have to fill this and also make
sure they will be packed in the packs if not at lease everyone else
knows how we intend to sell.