T 132 add ransfer of tax liability message to sk invoices with tax number
Would you be able to add this sentence on to all invoices for AT, CZ, EU which have VAT  Numbers on?

It can be on the bottom of the invoice somewhere. Check the picture Google has it too and we apparently need it.

SK: Prenesenie daňovej povinnosti
EU: Transfer of tax liability
AT:  Übertragung der Steuerschuld
CZ: Přenesení daňové povinnosti